+++ NEWS.md
@@ -12,6 +12,8 @@ Improvements
 ----------------
 
 * Updated object documentation to wrap lines longer than 80 characters. 
+* Added asMatrix() function to allow the pivot table contents to be retrieved as a matrix.
+* Added asDataFrame() function to allow the pivot table contents to be retrieved as a data frame.
 
 Bug Fixes
 ----------------